% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha}
\alias{alpha}
\title{Random Letter Generator}
\usage{
alpha(x = 1)
}
\arguments{
\item{x}{number of letters to return, defult set to x=1}
}
\value{
Random letter in alphabet
}
\description{
Function returns a random letter.
}
